+++ server/src/main/java/org/apache/iotdb/db/engine/compaction/level/LevelCompactionTsFileManagement.java
@@ -84,24 +84,28 @@ public class LevelCompactionTsFileManagement extends TsFileManagement {
     clear();
   }
 
-  private void deleteLevelFiles(long timePartitionId, Collection<TsFileResource> mergeTsFiles) {
-    logger.debug("{} [compaction] merge starts to delete file", storageGroupName);
+  private void deleteLevelFilesInDisk(Collection<TsFileResource> mergeTsFiles) {
+    logger.debug("{} [compaction] merge starts to delete real file", storageGroupName);
     for (TsFileResource mergeTsFile : mergeTsFiles) {
       deleteLevelFile(mergeTsFile);
       logger
           .info("{} [Compaction] delete TsFile {}", storageGroupName, mergeTsFile.getTsFilePath());
     }
-    for (int i = 0; i < seqLevelNum; i++) {
+  }
+
+  private void deleteLevelFilesInList(long timePartitionId,
+      Collection<TsFileResource> mergeTsFiles, int level, boolean sequence) {
+    logger.debug("{} [compaction] merge starts to delete file list", storageGroupName);
+    if (sequence) {
       if (sequenceTsFileResources.containsKey(timePartitionId)) {
-        if (sequenceTsFileResources.get(timePartitionId).size() > i) {
-          sequenceTsFileResources.get(timePartitionId).get(i).removeAll(mergeTsFiles);
+        if (sequenceTsFileResources.get(timePartitionId).size() > level) {
+          sequenceTsFileResources.get(timePartitionId).get(level).removeAll(mergeTsFiles);
         }
       }
-    }
-    for (int i = 0; i < unseqLevelNum; i++) {
+    } else {
       if (unSequenceTsFileResources.containsKey(timePartitionId)) {
-        if (unSequenceTsFileResources.get(timePartitionId).size() > i) {
-          unSequenceTsFileResources.get(timePartitionId).get(i).removeAll(mergeTsFiles);
+        if (unSequenceTsFileResources.get(timePartitionId).size() > level) {
+          unSequenceTsFileResources.get(timePartitionId).get(level).removeAll(mergeTsFiles);
         }
       }
     }
@@ -337,14 +341,20 @@ public class LevelCompactionTsFileManagement extends TsFileManagement {
                 .merge(targetTsFileResource, getTsFileList(isSeq), storageGroupName,
                     new CompactionLogger(storageGroupDir, storageGroupName), deviceSet, isSeq);
             if (isSeq) {
-              for (TreeSet<TsFileResource> currMergeFile : sequenceTsFileResources
-                  .get(timePartition)) {
-                deleteLevelFiles(timePartition, currMergeFile);
+              for (int level = 0; level < sequenceTsFileResources.get(timePartition).size();
+                  level++) {
+                TreeSet<TsFileResource> currLevelMergeFile = sequenceTsFileResources
+                    .get(timePartition).get(level);
+                deleteLevelFilesInDisk(currLevelMergeFile);
+                deleteLevelFilesInList(timePartition, currLevelMergeFile, level, isSeq);
               }
             } else {
-              for (List<TsFileResource> currMergeFile : unSequenceTsFileResources
-                  .get(timePartition)) {
-                deleteLevelFiles(timePartition, currMergeFile);
+              for (int level = 0; level < unSequenceTsFileResources.get(timePartition).size();
+                  level++) {
+                TreeSet<TsFileResource> currLevelMergeFile = sequenceTsFileResources
+                    .get(timePartition).get(level);
+                deleteLevelFilesInDisk(currLevelMergeFile);
+                deleteLevelFilesInList(timePartition, currLevelMergeFile, level, isSeq);
               }
             }
           }
@@ -370,13 +380,15 @@ public class LevelCompactionTsFileManagement extends TsFileManagement {
                 CompactionUtils
                     .merge(targetResource, sourceTsFileResources, storageGroupName,
                         new CompactionLogger(storageGroupDir, storageGroupName), deviceSet, true);
-                deleteLevelFiles(timePartition, sourceTsFileResources);
+                deleteLevelFilesInDisk(sourceTsFileResources);
+                deleteLevelFilesInList(timePartition, sourceTsFileResources, level, isSeq);
                 sequenceTsFileResources.get(timePartition).get(level + 1).add(targetResource);
               } else {
                 CompactionUtils
                     .merge(targetResource, sourceTsFileResources, storageGroupName,
                         new CompactionLogger(storageGroupDir, storageGroupName), deviceSet, false);
-                deleteLevelFiles(timePartition, sourceTsFileResources);
+                deleteLevelFilesInDisk(sourceTsFileResources);
+                deleteLevelFilesInList(timePartition, sourceTsFileResources, level, isSeq);
                 unSequenceTsFileResources.get(timePartition).get(level + 1).add(targetResource);
               }
             }
@@ -492,19 +504,20 @@ public class LevelCompactionTsFileManagement extends TsFileManagement {
                 storageGroupName, i, toMergeTsFiles.size());
             writeLock();
             try {
-              deleteLevelFiles(timePartition, toMergeTsFiles);
               compactionLogger.logMergeFinish();
               if (sequence) {
                 sequenceTsFileResources.get(timePartition).get(i + 1).add(newResource);
               } else {
                 unSequenceTsFileResources.get(timePartition).get(i + 1).add(newResource);
               }
+              deleteLevelFilesInList(timePartition, toMergeTsFiles, i, sequence);
               if (mergeResources.size() > i + 1) {
                 mergeResources.get(i + 1).add(newResource);
               }
             } finally {
               writeUnlock();
             }
+            deleteLevelFilesInDisk(toMergeTsFiles);
             compactionLogger.close();
             File logFile = FSFactoryProducer.getFSFactory()
                 .getFile(storageGroupDir, storageGroupName + COMPACTION_LOG_NAME);